Refactoring (#26)
